requirements.txt
django-q
Celery
djangorestframework
django